add custom emscripten sdl library
This version does not call preventDefault on canvas mousedown and mouseup
events. It seemed that preventDefault was stopping the canvas (and hence the
document) from gaining keyboard focus on click when in an iframe. This probably
isn't an issue when not in an iframe, because then the document probably already
has focus.